Use platform methods directly
