+++ server/channels/app/platform/metrics.go
@@ -25,6 +25,7 @@ import (
 	"github.com/mattermost/mattermost/server/public/model"
 	"github.com/mattermost/mattermost/server/public/plugin"
 	"github.com/mattermost/mattermost/server/public/shared/mlog"
+	"github.com/mattermost/mattermost/server/public/shared/request"
 	"github.com/mattermost/mattermost/server/v8/channels/utils"
 	"github.com/mattermost/mattermost/server/v8/einterfaces"
 )
@@ -253,7 +254,7 @@ func (ps *PlatformService) Metrics() einterfaces.MetricsInterface {
 	return ps.metricsIFace
 }
 
-func (ps *PlatformService) CreateCPUProfile() (*model.FileData, error) {
+func (ps *PlatformService) CreateCPUProfile(_ request.CTX) (*model.FileData, error) {
 	var b bytes.Buffer
 
 	err := rpprof.StartCPUProfile(&b)
@@ -272,7 +273,7 @@ func (ps *PlatformService) CreateCPUProfile() (*model.FileData, error) {
 	return fileData, nil
 }
 
-func (ps *PlatformService) CreateHeapProfile() (*model.FileData, error) {
+func (ps *PlatformService) CreateHeapProfile(_ request.CTX) (*model.FileData, error) {
 	var b bytes.Buffer
 
 	err := rpprof.Lookup("heap").WriteTo(&b, 0)
@@ -287,7 +288,7 @@ func (ps *PlatformService) CreateHeapProfile() (*model.FileData, error) {
 	return fileData, nil
 }
 
-func (ps *PlatformService) CreateGoroutineProfile() (*model.FileData, error) {
+func (ps *PlatformService) CreateGoroutineProfile(_ request.CTX) (*model.FileData, error) {
 	var b bytes.Buffer
 
 	err := rpprof.Lookup("goroutine").WriteTo(&b, 2)
