+++ server/channels/app/support_packet_test.go
@@ -326,96 +326,6 @@ func TestGenerateSupportPacket(t *testing.T) {
 	})
 }
 
-func TestGetNotificationsLog(t *testing.T) {
-	th := Setup(t)
-	defer th.TearDown()
-
-	// Disable notifications file setting in config so we should get an warning
-	th.App.UpdateConfig(func(cfg *model.Config) {
-		*cfg.NotificationLogSettings.EnableFile = false
-	})
-
-	fileData, err := th.App.getNotificationsLog(th.Context)
-	assert.Nil(t, fileData)
-	assert.ErrorContains(t, err, "Unable to retrieve notifications logs because NotificationLogSettings.EnableFile is set to false")
-
-	dir, err := os.MkdirTemp("", "")
-	require.NoError(t, err)
-	t.Cleanup(func() {
-		err = os.RemoveAll(dir)
-		assert.NoError(t, err)
-	})
-
-	// Enable notifications file but point to an empty directory to get an error trying to read the file
-	th.App.UpdateConfig(func(cfg *model.Config) {
-		*cfg.NotificationLogSettings.EnableFile = true
-		*cfg.LogSettings.FileLocation = dir
-	})
-
-	logLocation := config.GetNotificationsLogFileLocation(dir)
-
-	// There is no notifications.log file yet, so this fails
-	fileData, err = th.App.getNotificationsLog(th.Context)
-	assert.Nil(t, fileData)
-	assert.ErrorContains(t, err, "failed read notifcation log file at path "+logLocation)
-
-	// Happy path where we have file and no error
-	d1 := []byte("hello\ngo\n")
-	err = os.WriteFile(logLocation, d1, 0777)
-	require.NoError(t, err)
-
-	fileData, err = th.App.getNotificationsLog(th.Context)
-	assert.NoError(t, err)
-	require.NotNil(t, fileData)
-	assert.Equal(t, "notifications.log", fileData.Filename)
-	assert.Positive(t, len(fileData.Body))
-}
-
-func TestGetMattermostLog(t *testing.T) {
-	th := Setup(t)
-	defer th.TearDown()
-
-	// disable mattermost log file setting in config so we should get an warning
-	th.App.UpdateConfig(func(cfg *model.Config) {
-		*cfg.LogSettings.EnableFile = false
-	})
-
-	fileData, err := th.App.GetMattermostLog(th.Context)
-	assert.Nil(t, fileData)
-	assert.ErrorContains(t, err, "Unable to retrieve mattermost logs because LogSettings.EnableFile is set to false")
-
-	dir, err := os.MkdirTemp("", "")
-	require.NoError(t, err)
-	t.Cleanup(func() {
-		err = os.RemoveAll(dir)
-		assert.NoError(t, err)
-	})
-
-	// Enable log file but point to an empty directory to get an error trying to read the file
-	th.App.UpdateConfig(func(cfg *model.Config) {
-		*cfg.LogSettings.EnableFile = true
-		*cfg.LogSettings.FileLocation = dir
-	})
-
-	logLocation := config.GetLogFileLocation(dir)
-
-	// There is no mattermost.log file yet, so this fails
-	fileData, err = th.App.GetMattermostLog(th.Context)
-	assert.Nil(t, fileData)
-	assert.ErrorContains(t, err, "failed read mattermost log file at path "+logLocation)
-
-	// Happy path where we get a log file and no warning
-	d1 := []byte("hello\ngo\n")
-	err = os.WriteFile(logLocation, d1, 0777)
-	require.NoError(t, err)
-
-	fileData, err = th.App.GetMattermostLog(th.Context)
-	require.NoError(t, err)
-	require.NotNil(t, fileData)
-	assert.Equal(t, "mattermost.log", fileData.Filename)
-	assert.Positive(t, len(fileData.Body))
-}
-
 func TestCreateSanitizedConfigFile(t *testing.T) {
 	th := Setup(t)
 	defer th.TearDown()
